Addressing comments.

+++ python/PIK_maya_anim_sequencer/ui/windows.py
@@ -22,21 +22,48 @@ from quickBlast.main import run as quickBlast
 
 
 class Backend(QObject):
+    """
+    The backend object used in the QQuickWidget.
+    """
     def __init__(self):
         super().__init__()
         self.dialog_create_shot = None
 
     @Slot(str, result=str)
     def get_camera_name(self, shot_name):
+        """
+        Get the camera name for the given shot name.
+        Args:
+            shot_name (str): The shot name.
+
+        Returns:
+            str: The camera name.
+        """
         return shot_name
 
     @Slot(result=str)
     def get_shot_name(self):
+        """
+        Get the shot name.
+        Returns:
+            (str): The shot name.
+        """
         sequencer_sequence = get_sequencer_sequence()
         return sequencer_sequence.generate_shot_name()
 
     @Slot(str, str, int, str)
     def do_create_shot(self, shot_name, color, shot_length, shot_camera):
+        """
+        Create a shot at current time.
+        Args:
+            shot_name (str): The shot name.
+            color (str): The color of the shot as hex color code.
+            shot_length (int): The length of the shot.
+            shot_camera (str): The camera transform name of the shot.
+
+        Returns:
+            None
+        """
         sequencer_sequence = get_sequencer_sequence()
         shot = SequencerShot.create(shot_name, QColor(color).getRgbF(), shot_length)
         shot.cam = SequencerCamera.create(shot_camera)
@@ -49,12 +76,28 @@ class Backend(QObject):
 
     @Slot()
     def delete_shot(self):
+        """
+        Delete the shot at current time.
+        Returns:
+            None
+        """
         sequencer_sequence = get_sequencer_sequence()
         current_time = cmds.currentTime(query=True)
-        sequencer_sequence.delete_shot_at_time(current_time)
+        shot = sequencer_sequence.get_shots_at_time(current_time)[0]
+        if not shot:
+            OpenMaya.MGlobal.displayWarning("Sequencer: No active shot detected.")
+            return
+
+        if show_confirmation_dialog(f"Do you want to delete current shot?\n{shot.name}"):
+            sequencer_sequence.delete_shot_at_time(current_time)
 
     @Slot()
     def focus_active_shot(self):
+        """
+        Focus the active shot.
+        Returns:
+            None
+        """
         sequencer_sequence = get_sequencer_sequence()
         current_time = cmds.currentTime(query=True)
         active_shot = sequencer_sequence.get_shots_at_time(current_time)[0]
@@ -63,6 +106,11 @@ class Backend(QObject):
 
     @Slot()
     def defocus_active_shot(self):
+        """
+        Defocus the shot or expand the playback range.
+        Returns:
+            None
+        """
         sequencer_sequence = get_sequencer_sequence()
         current_time = cmds.currentTime(query=True)
         if (
@@ -76,40 +124,75 @@ class Backend(QObject):
 
     @Slot()
     def focus_previous_shot(self):
+        """
+        Focus the previous shot.
+        Returns:
+            None
+        """
         sequencer_sequence = get_sequencer_sequence()
         current_time = cmds.currentTime(query=True)
         sequencer_sequence.focus_previous_at_time(current_time)
 
     @Slot()
     def focus_next_shot(self):
+        """
+        Focus the next shot.
+        Returns:
+            None
+        """
         sequencer_sequence = get_sequencer_sequence()
         current_time = cmds.currentTime(query=True)
         sequencer_sequence.focus_next_at_time(current_time)
 
     @Slot()
     def reduce_active_shot_length(self):
+        """
+        Reduce active shot length by 1 time unit.
+        Returns:
+            None
+        """
         sequencer_sequence = get_sequencer_sequence()
         current_time = cmds.currentTime(query=True)
         sequencer_sequence.offset_frame_of_shot_at_time(current_time, -1)
 
     @Slot()
     def increase_active_shot_length(self):
+        """
+        Increase active shot length by 1 time unit.
+        Returns:
+            None
+        """
         sequencer_sequence = get_sequencer_sequence()
         current_time = cmds.currentTime(query=True)
         sequencer_sequence.offset_frame_of_shot_at_time(current_time, 1)
 
     @Slot()
     def link_shots(self):
+        """
+        Resolve gaps between shots.
+        Returns:
+            None
+        """
         sequencer_sequence = get_sequencer_sequence()
         sequencer_sequence.resolve_gaps_between_shots()
 
     @Slot()
     def unstack_shots(self):
+        """
+        Resolve overlapping shots.
+        Returns:
+            None
+        """
         sequencer_sequence = get_sequencer_sequence()
         sequencer_sequence.resolve_overlapping_shots()
 
     @Slot()
     def playblast(self):
+        """
+        Launch a playblast for each shot.
+        Returns:
+            None
+        """
         sequencer_sequence = get_sequencer_sequence()
 
         if cmds.file(query=True, modified=True):
@@ -167,34 +250,29 @@ class Backend(QObject):
         """
         sequencer_sequence = get_sequencer_sequence()
 
-        confirmation = cmds.confirmDialog(
-            title="Confirm",
-            message="Do you want to export the current sequence datas ?\nThis will export a datas.json file and a .ma file for each camera.",
-            button=["Yes", "No"],
-            defaultButton="No",
-            cancelButton="No",
-            dismissString="No",
-        )
-        if confirmation == "No":
-            return
+        if show_confirmation_dialog("Do you want to export the current sequence datas ?\nThis will export a datas.json file and a .ma file for each camera."):
+            folder_path = cmds.fileDialog2(fileMode=2, dialogStyle=1, hideNameEdit=True)
 
-        folder_path = cmds.fileDialog2(fileMode=2, dialogStyle=1, hideNameEdit=True)
+            if folder_path:
+                folder_path = folder_path[0]
 
-        if folder_path:
-            folder_path = folder_path[0]
+                datas = list()
+                for shot in sequencer_sequence.shots:
+                    datas.append(shot.as_dict())
+                    shot.export_camera(folder_path)
 
-            datas = list()
-            for shot in sequencer_sequence.shots:
-                datas.append(shot.as_dict())
-                shot.export_camera(folder_path)
+                with open(os.path.join(folder_path, "datas.json"), "w") as file:
+                    json.dump(datas, file, indent=4)
 
-            with open(os.path.join(folder_path, "datas.json"), "w") as file:
-                json.dump(datas, file, indent=4)
-
-        os.startfile(folder_path)
+            os.startfile(folder_path)
 
     @Slot()
     def open_create_shot_dialog(self):
+        """
+        Show a modal dialog to create a new shot.
+        Returns:
+            None
+        """
         folderpath = os.path.dirname(__file__)
         qml_url = os.path.join(folderpath, "dialog_createShot.qml")
 
@@ -213,6 +291,11 @@ class Backend(QObject):
 
     @Slot()
     def close_create_shot_dialog(self):
+        """
+        Close the create shot dialog window.
+        Returns:
+            None
+        """
         if self.dialog_create_shot:
             self.dialog_create_shot.close()
 
@@ -221,6 +304,9 @@ backend = Backend()
 
 
 class DockableMainWindow(MayaQWidgetDockableMixin, QMainWindow):
+    """
+    The Sequencer main window object.
+    """
     def __init__(self, qml_url: str, width: int, height: int, *args, **kwargs):
         super(DockableMainWindow, self).__init__(*args, **kwargs)
 
@@ -236,6 +322,9 @@ class DockableMainWindow(MayaQWidgetDockableMixin, QMainWindow):
 
 
 class DialogCreateShotWindow(QDialog):
+    """
+    The create shot dialog window object.
+    """
     def __init__(self, qml_url: str, width: int, height: int, *args, **kwargs):
         super(DialogCreateShotWindow, self).__init__(*args, **kwargs)
 
@@ -254,6 +343,107 @@ class DialogCreateShotWindow(QDialog):
         self.setModal(True)
 
 
+def show_confirmation_dialog(text) -> bool:
+    """
+    Show a confirmation dialog. Return True if the dialog was successful.
+    Args:
+        text (str): The message to display.
+
+    Returns:
+        bool: True if user confirmed successfully.
+    """
+    confirmation = cmds.confirmDialog(
+        title="Confirm",
+        message=text,
+        button=["Yes", "No"],
+        defaultButton="No",
+        cancelButton="No",
+        dismissString="No",
+    )
+    if confirmation == "Yes":
+        return True
+    else:
+        return False
+
+
+def create_viewport(label: str, viewport_size: tuple):
+    """
+    Create a floating viewport and return the model panel name.
+    Args:
+        label (str): The label of the viewport.
+        viewport_size (tuple): The size of the viewport.
+
+    Returns:
+        str: The name of the panel containing the viewport.
+    """
+    dock_name = "PIK_" + label.title().replace(" ", "") + "Dock"
+
+    if cmds.workspaceControl(dock_name + "WorkspaceControl", exists=True):
+        cmds.workspaceControl(dock_name + "WorkspaceControl", edit=True, close=True)
+
+    workspace = cmds.workspaceControl(
+        dock_name + "WorkspaceControl",
+        label=label,
+        floating=True,
+        initialWidth=viewport_size[0],
+        initialHeight=viewport_size[1],
+        retain=False,
+    )
+
+    cmds.setParent(workspace)
+
+    cmds.paneLayout()
+    panel = cmds.modelPanel()
+
+    # Enable shading, manage visibilities
+    cmds.modelEditor(
+        panel,
+        edit=True,
+        displayAppearance="smoothShaded",
+        wireframeOnShaded=False,
+        grid=False,
+        nurbsCurves=False,
+        locators=False,
+        cameras=False,
+        joints=False,
+        imagePlane=False,
+        follicles=False,
+        displayTextures=True,
+    )
+
+    return panel
+
+
 def get_maya_main_window():
+    """
+    Get the maya main window object.
+    Returns:
+        :class: `QWidget`: The maya main window object.
+    """
     main_window_ptr = OpenMayaUI.MQtUtil.mainWindow()
     return wrapInstance(int(main_window_ptr), QWidget)
+
+
+def maya_dock_control_to_window(window: str, to_dock: str, position: str = "bottom"):
+    """
+    Dock a maya control to a maya window.
+    Args:
+        window (str): The name of the workspace control to dock to.
+        to_dock (str): The name of the control to dock.
+        position (str): The position of the control to dock (valid values are: "left", "right", "top", "bottom"). Defaults to bottom.
+
+    Returns:
+        None
+    """
+    try:
+        control_to_dock = cmds.modelPanel(
+            to_dock, query=True, control=True
+        )
+        if control_to_dock:
+            control_to_dock = control_to_dock.split("|")[0]
+            cmds.workspaceControl(
+                window, edit=True, dockToControl=(control_to_dock, position)
+            )
+    except RuntimeError:
+        # If the viewport is already docked, Maya will raise an error.
+        pass
